+++ kahypar/partition/initial_partitioning/greedy_hypergraph_growing_initial_partitioner.h
@@ -281,7 +281,7 @@ class GreedyHypergraphGrowingInitialPartitioner : public IInitialPartitioner,
     // TODO(heuer): Shouldn't it be possible to do this within the deltaGainUpdate function?
     if (insert) {
       for (const HyperedgeID& he : _hg.incidentEdges(hn)) {
-        if (!_hyperedge_in_queue[target_part * _hg.initialNumEdges() + he]) {
+        if (!_hyperedge_in_queue[static_cast<size_t>(target_part) * _hg.initialNumEdges() + he]) {
           if (_hg.edgeSize(he) <= _config.partition.hyperedge_size_threshold) {
             for (const HypernodeID& pin : _hg.pins(he)) {
               if (_hg.partID(pin) == _config.initial_partitioning.unassigned_part) {
@@ -292,7 +292,8 @@ class GreedyHypergraphGrowingInitialPartitioner : public IInitialPartitioner,
               }
             }
           }
-          _hyperedge_in_queue.set(target_part * _hg.initialNumEdges() + he, true);
+          _hyperedge_in_queue.set(static_cast<size_t>(target_part) * _hg.initialNumEdges() + he,
+                                  true);
         }
       }
     }
